Fix: bd-key-value 32-bit Android compilation

+++ bd-key-value/src/lib.rs
@@ -168,7 +168,7 @@ impl Store {
               anyhow::bail!("stored string is too short to contain length prefix")
             };
 
-            let len = usize::from_le_bytes(*len_bytes);
+            let len = u64::from_le_bytes(*len_bytes) as usize;
 
             if rest.len() != len {
               anyhow::bail!("stored string length prefix does not match actual length");
